feat: BadgeViewModel 분리
